backend/requirements.txt
# SmartRiver Backend — Python dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data & ML
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.14.0

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
